pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "credweaver"
version = "2.0.0"
description = "CredWeaver — Weave profile data into targeted credential wordlists. Python + Rust hybrid for authorized pentesting."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "CredWeaver Contributors" }]
keywords = ["security", "pentesting", "wordlist", "password", "credentials", "osint", "ctf", "cracking"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Rust",
    "Topic :: Security",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "typer[all]>=0.12",
    "rich>=13",
    "pydantic>=2",
    "pyyaml>=6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.4",
    "mypy>=1.10",
    "maturin>=1.7",
    "pre-commit>=3.7",
]

[project.scripts]
credweaver = "credweaver.cli:main"
cw = "credweaver.cli:main"

[tool.maturin]
module-name = "credweaver_engine"
manifest-path = "rust_engine/Cargo.toml"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "C4", "SIM"]
ignore = [
    "E501",   # line too long — handled by formatter
    "SIM108", # ternary operator — not always clearer
]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
